Final implementation.
